Fixed huge bug about Timer thread not terminating
Specifically, the scheduled TimerTask to display the next question was
not being terminated when you closed the program. This happened when you
closed the program after clicking an answer but before the next question
was displayed. This caused the program / Java to stay open in the
background until you force closed it with something like Task Manager.
Now added a setOnCloseRequest to completely terminate all threads.